+++ R/readers.R
@@ -1,4 +1,4 @@
-#' Reads PK datasets from various file formats.
+#' Reads datasets from various file formats.
 #'
 #' @details
 #' Currently supported file formats include:
@@ -8,7 +8,8 @@
 #' @returns A data.frame object with loaded data.
 #'
 #' @examples
-#' df <- read_pk(system.file("shiny/data/example-ADNCA.csv", package = "aNCA"))
+#' path <- system.file("shiny/tests/testthat/dummy_simplified.csv", package = "aNCA")
+#' df <- read_pk(path)
 #'
 #' @importFrom tools file_ext
 #' @importFrom utils read.csv
@@ -53,15 +54,15 @@ readers <- list(
   csv = function(path) read.csv(path, na = c("", "NA")),
   rds = function(path) readRDS(path),
   xlsx = function(path) {
-    if (!requireNamespace("openxlsx2", silently = TRUE))
+    if (!requireNamespace("readxl", quietly = TRUE))
       stop(
-        "Handling .xlsx files requires `openxlsx2` package, please install it with ",
-        "`install.packages('openxlsx2')`"
+        "Handling .xlsx files requires `readxl` package, please install it with ",
+        "`install.packages('readxl')`"
       )
-    openxlsx2::read_xlsx(path)
+    as.data.frame(readxl::read_excel(path))
   },
   sas7bdat = function(path) {
-    if (!requireNamespace("haven", silently = TRUE))
+    if (!requireNamespace("haven", quietly = TRUE))
       stop(
         "Handling .sas7bdat files requires `haven` package, please install it with ",
         "`install.packages('haven')`"
@@ -69,7 +70,7 @@ readers <- list(
     haven::read_sas(path)
   },
   xpt = function(path) {
-    if (!requireNamespace("haven", silently = TRUE))
+    if (!requireNamespace("haven", quietly = TRUE))
       stop(
         "Handling .xpt files requires `haven` package, please install it with ",
         "`install.packages('haven')`"
@@ -77,7 +78,7 @@ readers <- list(
     haven::read_xpt(path)
   },
   parquet = function(path) {
-    if (!requireNamespace("arrow", silently = TRUE))
+    if (!requireNamespace("arrow", quietly = TRUE))
       stop(
         "Handling .parquet files requires `arrow` package, please install it with ",
         "`install.packages('arrow')`"