+++ inst/www/templates/script_template.R
@@ -8,7 +8,7 @@ library(aNCA)
 library(dplyr)
 
 # Load raw data #
-data_path <- "../data/data.rds"
+data_path <- "../input_data.rds"
 adnca_data <- read_pk(data_path)
 
 ## Preprocess data ########################################
@@ -17,9 +17,6 @@ names(mapping) <- gsub("select_", "", names(mapping))
 applied_filters <- settings_list$applied_filters
 
 preprocessed_adnca <- adnca_data %>%
-
-  # Filter the data
-  apply_filters(applied_filters) %>%
 
   # Map columns to their standards
   apply_mapping(
@@ -37,15 +34,18 @@ preprocessed_adnca <- adnca_data %>%
   create_metabfl(mapping$Metabolites) %>%
   
   # Make sure all variables are in its correct class
-  adjust_class_and_length(metadata_nca_variables)
+  adjust_class_and_length(metadata_nca_variables) %>%
+
+  # Filter the data
+  apply_filters(applied_filters)
 
 ## Setup NCA settings in the PKNCA object ########################
 int_parameters <- settings_list$settings$int_parameters
-units_table <- settings_list$final_units
+units_table <- settings_list$units_table
 parameters_selected_per_study <- settings_list$settings$parameters$selections
 study_types_df <- settings_list$settings$parameters$types_df
 extra_vars_to_keep <-  settings_list$extra_vars_to_keep
-slope_rules <- settings_list$slope_rules$manual_slopes
+slope_rules <- settings_list$slope_rules
 
 pknca_obj <- preprocessed_adnca %>%
 
@@ -77,7 +77,12 @@ pknca_obj <- preprocessed_adnca %>%
   {
     pknca_obj <- .
     if (!is.null(units_table)) {
-      pknca_obj[["units"]] <- units_table
+      pknca_obj[["units"]] <- dplyr::rows_update(
+        pknca_obj[["units"]],
+        units_table,
+        by = c("PPTESTCD", "PPORRESU"),
+        unmatched = "ignore"
+      )
     }
     pknca_obj
   }